Update gyro_and_motors.ino
added rudimentary turning and forward/backward drive. Driving forward/backward doesn't currently work, because robot tries to maintain a slight tilt, which leads to perpetum acceleration, and at one point, motors are going so fast that they are capped at one end.
Gonna try a fix, where robot tries to instead get the motor's speed up to a certain point, but when that point is reached, the robot will try to balance itself on the standing still point, instead of that tilt point.