Update gyro_and_motors.ino
Now uses a joystick drive with an android app. Todo: Make it more stable and add a method to recognize when it is going too fast to keep up and take measures to prevent it (either by tuning pid or by detecting motors speed and setting a new zero reference point whenever motors are going too fast)

TL;DR: tune the settings so that the robot will never topple over in normal settings or with slight nudging.